Fix tests

+++ gammapy/maps/regionnd.py
@@ -347,7 +347,8 @@ class RegionNDMap(Map):
             Array to be used as weights. The spatial geometry must be equivalent
             to `other` and additional axes must be broadcastable.
         """
-        data = other.quantity.to_value(self.unit)
+        # TODO: check and remove the type-cast here...
+        data = other.quantity.to_value(self.unit).astype(self.data.dtype)
 
         # TODO: re-think stacking of regions. Is making the union reasonable?
         # self.geom.union(other.geom)
